pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "factstack"
version = "0.1.0"
description = "Evidence-first RAG Q&A system for technical documentation"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "FactStack Team"}
]
dependencies = [
    "chromadb>=0.4.0",
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "rank-bm25>=0.2.2",
    "tiktoken>=0.5.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
factstack-ingest = "factstack.ingest:main"
factstack-ask = "factstack.ask:main"
factstack-eval = "factstack.eval:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
factstack = ["py.typed"]
